requirements.txt
discord.py
python-dotenv~=0.20.0
Pillow
numpy
asyncio
PyNaCl